Separated files for Settings & Rendering
